# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cairn API",
    "description": "\nCairn walks a family through the logistics of a death, one step at a time.\n\nEvery response that moves a conversation forward carries a `next_step` with a\nsingle prompt, so clients can ask one question at a time. Errors use RFC 9457\nproblem details (`application/problem+json`) and never echo submitted values.\n\nAuthentication is an Auth0 access token. Accounts can be created with Google,\nApple, or an email magic link (see GET /v1/welcome). Onboarding follows\nUC-REG-01 to UC-REG-14, one question per screen. Case creation follows\nUC-CASE-01 to UC-CASE-18: a case starts as a draft, and the free period of 28\ndays starts only when the user taps Start journey on their first case. After\nit ends the account is read-only until subscribed. No payment information is\nasked for anywhere in case creation. Free text is redacted as it is received\nand never stored. The case is the security boundary and is enforced in the\ndatabase with row-level security.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/cases": {
      "get": {
        "tags": [
          "Case creation"
        ],
        "summary": "Your cases, drafts included",
        "operationId": "list_cases_v1_cases_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Case creation"
        ],
        "summary": "Start a new case",
        "description": "UC-CASE-01 and UC-CASE-18. Creates a draft case right away and acknowledges the loss before asking anything. Offers two ways in: one question at a time, or in your own words. Creating or editing a draft never starts the free period (DEC-01). A read-only account can create a draft. Nothing about the person who died is collected here.",
        "operationId": "create_case_v1_cases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/StartCaseRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeTurnResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of this case, the case does not exist, or (for an active case) the account is read-only (code account_read_only)."
          },
          "409": {
            "description": "Onboarding isn't finished, or a changed acknowledgment needs to be agreed to."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/cases/{case_id}": {
      "get": {
        "tags": [
          "Case creation"
        ],
        "summary": "Open a case",
        "description": "UC-CASE-10. Opening a draft counts as activity and resets its 28 days. The response greets the user, says in one line where they left off, and asks whether to continue.",
        "operationId": "get_case_v1_cases__case_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of this case, the case does not exist, or (for an active case) the account is read-only (code account_read_only)."
          },
          "409": {
            "description": "Onboarding isn't finished, or a changed acknowledgment needs to be agreed to."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}/deceased": {
      "patch": {
        "tags": [
          "Case creation"
        ],
        "summary": "Record the deceased's legal identity, inside the task that needs it",
        "description": "Just in time only. full_legal_name and date_of_birth are never part of case creation (never_collect_at_case_creation). Refused while the case is a draft. The database refuses it too.",
        "operationId": "patch_deceased_v1_cases__case_id__deceased_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeceasedIdentityPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a member of this case, the case does not exist, or (for an active case) the account is read-only (code account_read_only)."
          },
          "409": {
            "description": "The case is still a draft."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaseListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "last_activity_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Activity At"
          },
          "draft_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "display_name",
          "last_activity_at",
          "draft_expires_at"
        ],
        "title": "CaseListItem"
      },
      "CaseListResponse": {
        "properties": {
          "cases": {
            "items": {
              "$ref": "#/components/schemas/CaseListItem"
            },
            "type": "array",
            "title": "Cases"
          }
        },
        "type": "object",
        "required": [
          "cases"
        ],
        "title": "CaseListResponse"
      },
      "CaseOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "read_only",
              "paused",
              "closed"
            ],
            "title": "Status",
            "description": "read_only when the case is active and the account's free period has ended."
          },
          "display_name": {
            "type": "string",
            "title": "Display Name",
            "description": "What to call the person who died. Not a legal name."
          },
          "journey_template_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Journey Template Key"
          },
          "journey_template_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Journey Template Version"
          },
          "journey_started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Journey Started At"
          },
          "journey_started_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Journey Started On"
          },
          "last_intake_step": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FieldKey"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_activity_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Activity At"
          },
          "draft_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Expires At",
            "description": "Drafts only. Any activity moves it out again."
          },
          "death_not_yet_occurred": {
            "type": "boolean",
            "title": "Death Not Yet Occurred"
          },
          "skip_explainers": {
            "type": "boolean",
            "title": "Skip Explainers"
          },
          "tasks_paused_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tasks Paused Until"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "display_name",
          "journey_template_key",
          "journey_template_version",
          "journey_started_at",
          "journey_started_on",
          "last_intake_step",
          "last_activity_at",
          "draft_expires_at",
          "death_not_yet_occurred",
          "skip_explainers",
          "tasks_paused_until",
          "created_at"
        ],
        "title": "CaseOut"
      },
      "CaseResponse": {
        "properties": {
          "case": {
            "$ref": "#/components/schemas/CaseOut"
          },
          "deceased": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeceasedOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "Only once a task has collected it."
          },
          "acknowledgment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledgment"
          },
          "body": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Body"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/Note"
            },
            "type": "array",
            "title": "Notes"
          },
          "next_step": {
            "$ref": "#/components/schemas/NextStep"
          },
          "read_aloud": {
            "$ref": "#/components/schemas/ReadAloud"
          }
        },
        "type": "object",
        "required": [
          "case",
          "deceased",
          "notes",
          "next_step",
          "read_aloud"
        ],
        "title": "CaseResponse",
        "description": "Opening a case. For a draft this is the resume turn (UC-CASE-10) and counts as activity."
      },
...
      "DeceasedIdentityPatch": {
        "properties": {
          "legal_first_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal First Name"
          },
          "legal_middle_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Middle Name"
          },
          "legal_last_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Last Name"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DeceasedIdentityPatch",
        "description": "Just-in-time identity, inside a task that needs it, once the journey has started.\n\nNever part of case creation. Omitted fields are left unchanged. Null clears a field."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IntakeSession": {
        "properties": {
          "safety_mode": {
            "$ref": "#/components/schemas/SafetyMode",
            "default": "normal"
          },
          "sensitivity": {
            "type": "string",
            "enum": [
              "normal",
              "raised"
            ],
            "title": "Sensitivity",
            "default": "normal"
          },
          "consecutive_skips": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Consecutive Skips",
            "default": 0
          },
          "offered": {
            "items": {
              "type": "string",
              "const": "loss_survivor_resources"
            },
            "type": "array",
            "maxItems": 4,
            "title": "Offered"
          },
          "ask_residence": {
            "type": "boolean",
            "title": "Ask Residence",
            "description": "The user said the death happened away from home, so residence_state is asked (UC-CASE-04).",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "IntakeSession",
        "description": "Per-session state the client keeps and sends back with each turn (UC-CASE-14).\n\nThe server never stores it. Decision 7 in database/CLAUDE.md: no inference\nabout the user's emotional state is persisted. Send back the session from\nthe last response. Omit it at the start of a new session."
      },
      "IntakeTurnResponse": {
        "properties": {
          "case": {
            "$ref": "#/components/schemas/CaseOut"
          },
          "voice": {
            "type": "string",
            "enum": [
              "steady_direct",
              "warm_patient",
              "brisk_businesslike",
              "plain_practical",
              "steady_care"
            ],
            "title": "Voice"
          },
          "safety_mode": {
            "$ref": "#/components/schemas/SafetyMode"
          },
          "acknowledgment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledgment"
          },
          "body": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Body",
            "description": "Statements, never questions."
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/Note"
            },
            "type": "array",
            "title": "Notes"
          },
          "question": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Question"
              },
              {
                "type": "null"
              }
            ]
          },
          "proposals": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProposedAnswer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposals",
            "description": "Free-text readback. Nothing is saved until confirmed."
          },
          "support": {
            "items": {
              "$ref": "#/components/schemas/SupportResource"
            },
            "type": "array",
            "title": "Support"
          },
          "redactions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Redactions",
            "description": "Kinds of values removed, never the values."
          },
          "masked_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Masked Text",
            "description": "The user's message with sensitive numbers removed. Replace the local copy with this."
          },
          "next_step": {
            "$ref": "#/components/schemas/NextStep"
          },
          "session": {
            "$ref": "#/components/schemas/IntakeSession"
          },
          "read_aloud": {
            "$ref": "#/components/schemas/ReadAloud"
          }
        },
        "type": "object",
        "required": [
          "case",
          "voice",
          "safety_mode",
          "next_step",
          "session",
          "read_aloud"
        ],
        "title": "IntakeTurnResponse",
        "description": "One conversational turn. Acknowledgment first, at most one question, exactly one next action."
      },
...
      "NextStep": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action",
            "description": "Machine-readable step code the client routes on."
          },
          "prompt": {
            "type": "string",
            "title": "Prompt",
            "description": "Plain-language text to show the user."
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Option"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "action",
          "prompt"
        ],
        "title": "NextStep",
        "description": "The single next thing to ask or do. Clients render one question at a time."
      },
      "Note": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "acknowledgment",
              "info",
              "legal",
              "crisis",
              "reminder",
              "account"
            ],
            "title": "Kind"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "legal_review_required": {
            "type": "boolean",
            "title": "Legal Review Required",
            "default": false
          },
          "source_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Urls",
            "description": "Citations. Render next to the text."
          },
          "attorney_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attorney Line",
            "description": "Shown with anything flagged for an attorney."
          }
        },
        "type": "object",
        "required": [
          "kind",
          "text"
        ],
        "title": "Note"
      },
...
      "ProposedAnswer": {
        "properties": {
          "field": {
            "$ref": "#/components/schemas/FieldKey"
          },
          "value": {
            "title": "Value"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "How the value reads back to the user."
          },
          "own_words": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Own Words"
          }
        },
        "type": "object",
        "required": [
          "field",
          "value",
          "label"
        ],
        "title": "ProposedAnswer"
      },
...
      "Question": {
        "properties": {
          "field": {
            "$ref": "#/components/schemas/FieldKey"
          },
          "pre_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pre Question",
            "description": "Why Cairn asks. Shown before the question."
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "input": {
            "type": "string",
            "enum": [
              "choice",
              "multi_choice",
              "text",
              "date_of_death",
              "place_of_death",
              "residence_state"
            ],
            "title": "Input"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/AnswerOption"
            },
            "type": "array",
            "title": "Options",
            "description": "Answer choices. Large tap targets."
          },
          "skip": {
            "$ref": "#/components/schemas/AnswerOption",
            "description": "Skip for now. On every question."
          },
          "not_sure": {
            "$ref": "#/components/schemas/AnswerOption",
            "description": "I'm not sure. On every question."
          },
          "free_text_allowed": {
            "type": "boolean",
            "title": "Free Text Allowed",
            "default": true
          },
          "free_text_label": {
            "type": "string",
            "title": "Free Text Label"
          }
        },
        "type": "object",
        "required": [
          "field",
          "prompt",
          "input",
          "options",
          "skip",
          "not_sure",
          "free_text_label"
        ],
        "title": "Question"
      },
...
      "SafetyMode": {
        "type": "string",
        "enum": [
          "normal",
          "overwhelm",
          "acute_distress",
          "risk_of_harm"
        ],
        "title": "SafetyMode"
      },
...
      "StartCaseRequest": {
        "properties": {
          "user_role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              },
              {
                "type": "null"
              }
            ],
            "description": "The answer from the onboarding hand-off (CaseHandoffResponse.user_role), so the question is not asked twice. Omit to ask it during intake."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "StartCaseRequest"
      },
...
      "SupportResource": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "text",
          "url"
        ],
        "title": "SupportResource"
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "spouse_partner",
          "child",
          "other_family",
          "named_executor",
          "power_of_attorney",
          "professional_fiduciary",
          "friend",
          "other",
          "prefer_not_to_say"
        ],
        "title": "UserRole"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Auth0 access token for the Cairn API audience.",
        "scheme": "bearer"
      }
    }
  },
...
}